Full submission up to the start of general info
